# Conflicts:
#	data/src/main/java/com/yapp/bol/data/remote/GroupApi.kt

+++ presentation/src/main/java/com/yapp/bol/presentation/viewmodel/login/MyGroupList.kt
@@ -13,7 +13,7 @@ object MyGroupList {
         return myGroupList
     }
 
-    fun findMyGroup(groupId: Int?): JoinedGroupItem? {
-        return myGroupList?.find { it.id.toInt() == groupId }
+    fun findMyGroup(groupId: Long?): JoinedGroupItem? {
+        return myGroupList?.find { it.id == groupId }
     }
 }
